+++ EssentialFeed/EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -1,0 +1,37 @@
+//
+//  FeedItemsMapper.swift
+//  EssentialFeed
+//
+//  Created by Arpad Zalan on 2024. 09. 24..
+//
+
+import Foundation
+
+final class FeedItemsMapper {
+    private struct Root: Decodable {
+        let items: [Item]
+    }
+
+    private struct Item: Decodable {
+        let id: UUID
+        let description: String?
+        let location: String?
+        let image: URL
+        
+        var item: FeedItem {
+            FeedItem(id: id, description: description, location: location, imageURL: image)
+        }
+    }
+    
+    private static var OK_200: Int { 200 }
+
+    static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [FeedItem] {
+        guard response.statusCode == OK_200 else {
+            throw RemoteFeedLoader.Error.invalidData
+        }
+        return try JSONDecoder()
+            .decode(Root.self, from: data)
+            .items
+            .map { $0.item }
+    }
+}
